cuarto commit

+++ src/controllers/cart.controllers.js
@@ -1,13 +1,34 @@
 const catchError = require("../utils/catchError");
 const Cart = require("../models/Cart");
+const Product = require("../models/Product");
+const Category = require("../models/Category");
 
 const getAll = catchError(async (req, res) => {
-  const results = await Cart.findAll();
+  const userId = req.user.id;
+  const results = await Cart.findAll({
+    where: { userId },
+    include: [
+      {
+        model: Product,
+        attributes: { exclude: ["createdAt","updatedAt"] }
+        include: [
+          {
+            model: Category,
+            attributes: [ 'name']
+          }
+        ],
+      },
+    ],
+  });
   return res.json(results);
 });
 
 const create = catchError(async (req, res) => {
-  const result = await Cart.create(req.body);
+  const { quantity, productId } = req.body;
+  const userId = req.user.id;
+
+  const body = { userId, quantity, productId };
+  const result = await Cart.create(body);
   return res.status(201).json(result);
 });
 
@@ -20,17 +41,19 @@ const getOne = catchError(async (req, res) => {
 
 const remove = catchError(async (req, res) => {
   const { id } = req.params;
-  const result = await Cart.destroy({ where: { id } });
+  const result = await Cart.destroy({ where: { id, userId: req.user.id } });
   if (!result) return res.sendStatus(404);
   return res.sendStatus(204);
 });
 
 const update = catchError(async (req, res) => {
+  const userId = req.user.id;
   const { id } = req.params;
-  const result = await Cart.update(req.body, {
-    where: { id },
-    returning: true,
-  });
+  const { quantity } = req.body;
+  const result = await Cart.update(
+    { quantity },
+    { where: { id, userId }, returning: true }
+  );
   if (result[0] === 0) return res.sendStatus(404);
   return res.json(result[1][0]);
 });
